Permit deletedValue option
I haven't tested this change very thoroughly but this should be backwards compatible. It opens the option to specify the values to be assigned to a deleted row. This modification permits me to use a `deletedAt` timestamp instead of a simple boolean. This information is often valuable to me. Here's the options I use to produce this functionality:

```javascript
const knex = Knex({ dialect: 'pg' })
{
  columnName: 'deletedAt',
  deletedValue: knex.fn.now(),
  notDeletedValue: null
}
```